`Update Jenkinsfile Test and SonarQube stages`
Modified the `Test (JUnit / Jacoco)` and `SonarQube analysis` stages in the Jenkinsfile. In the `Test (JUnit / Jacoco)` stage, removed the command to run Gradle tests and generate a Jacoco test report, and replaced it with a script to print "Run Unit Tests". Also removed the `junit` command to publish the test results. In the `SonarQube analysis` stage, replaced the command to run the SonarQube analysis using Gradle with a script that prints "Run SonarQube Scanner".