`Fix: Correct Trivy Scan command interpretation in Jenkinsfile`
` `

`The Jenkinsfile has been updated to correct the interpretation of the Trivy Scan command. Previously, the command was not being interpreted as a string, leading to execution errors. This has been fixed by enclosing the command in double quotes. This command is used to run a Docker container with the Trivy scanner, scanning the 'cc-backend' image. The image version is determined by the build number from the Jenkins environment variable.`