Add 'Deploy to ECR' stage to Jenkins pipeline
This commit introduces a new stage in the Jenkins pipeline named 'Deploy to ECR'. It encompasses steps that execute a script for determining the branch name by accessing the first branch in the SCM's branches list. Furthermore, it sets up Docker to interact with a registry by constructing the registry URL from the `ECR_FQDN` environment variable. Authentication is achieved using credentials based on the `DEPLOY_REGION` environment variable, the AWS service identifier `ecr`, and the Jenkins credentials ID `aws-rapid-jenkins-ecr-user`.