Update Docker image name in Jenkinsfile
This commit updates the Docker image name from `cc-cases` to `cc-users` in the Jenkinsfile. The change affects both the `Build` stage, where the `docker.build` command's argument is modified to include the new image name and build number, and the `Trivy Image Scan` stage, ensuring the Trivy scanner checks the newly named image for vulnerabilities.